test: split flow lifecycle coverage by topic
